magnecruit_backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-SocketIO
Flask-Cors
python-dotenv
psycopg2-binary
requests      
google.generativeai
eventlet
python-dotenv
langchain
langchain-community
crewai